Use a total_bytes for byte counts inside raptor_www

+++ src/raptor_internal.h
@@ -933,7 +933,7 @@ struct  raptor_www_s {
   raptor_world* world;
   char *type;
   int free_type;
-  int total_bytes;
+  size_t total_bytes;
   int failed;
   int status_code;
